Fix error when deleting users without avatar

+++ admin/src/Controllers/Users.php
@@ -236,7 +236,7 @@ class Users extends AbstractController
     protected function deleteAvatar(User $user)
     {
         $avatar = $user->avatar()->path();
-        if (FileSystem::exists($avatar)) {
+        if ($avatar !== null && FileSystem::exists($avatar)) {
             FileSystem::delete($avatar);
         }
     }
